+++ RELEASE_NOTES.md
@@ -51,6 +51,7 @@ What's New
 * #969 - correct LinesOfText in DDs to not be required; make ATs not auto-echo News (gbirchmeier)
 * #679 - Change logging to use .NET ILogger API aka Microsoft.Extensions.Logging (jkulubya/gbirchmeier)
          The existing logging interface is preserved and can still be used.
+* #965 - Reusing StringBuilder with Object Pooling (VAllens)
 
 ### v1.13.1
 * backport #951 to 1.13